func tests: helpers.remote: improve commands' streams handling
First of all, do not read remote streams sequentially: this will
deadlock if the remote's output is interleaved on both streams and the
second stream fills its buffer.

Next, improve situation with stderr. The `ignore_stderr` flag to
helpers.remote.Node.run_cmd() is only used to counter a problem of
Nginx: it does not release stderr after daemonization, thus blocking
communicate() indefinitely. Try to recover it even in this situation,
basically reimplementing Popen.communicate() so that it specifically
doesn't wait for EOF on the stderr stream.